+++ plugins/woocommerce-admin/packages/navigation/src/test/index.js
@@ -2,7 +2,7 @@
 /**
  * Internal dependencies
  */
-import { getPersistedQuery } from '../index';
+import { getPersistedQuery, getSearchWords } from '../index';
 
 jest.mock( '../index', () => ( {
 	...require.requireActual( '../index' ),
@@ -14,6 +14,7 @@ jest.mock( '../index', () => ( {
 		after: '2018-02-01',
 		before: '2018-01-01',
 		interval: 'day',
+		search: 'lorem',
 	} ),
 } ) );
 
@@ -61,5 +62,50 @@ describe( 'getPersistedQuery', () => {
 		};
 
 		expect( getPersistedQuery() ).toEqual( persistedQuery );
+	} );
+} );
+
+describe( 'getSearchWords', () => {
+	it( 'should get the search words from a query object', () => {
+		const query = {
+			search: 'lorem,dolor sit',
+		};
+		const searchWords = [ 'lorem', 'dolor sit' ];
+
+		expect( getSearchWords( query ) ).toEqual( searchWords );
+	} );
+
+	it( 'should parse `%2C` as commas', () => {
+		const query = {
+			search: 'lorem%2Cipsum,dolor sit',
+		};
+		const searchWords = [ 'lorem,ipsum', 'dolor sit' ];
+
+		expect( getSearchWords( query ) ).toEqual( searchWords );
+	} );
+
+	it( 'should return an empty array if the query has no `search` property', () => {
+		const query = {};
+		const searchWords = [];
+
+		expect( getSearchWords( query ) ).toEqual( searchWords );
+	} );
+
+	it( 'should use the persisted query when it receives no params', () => {
+		const searchWords = [ 'lorem' ];
+
+		expect( getSearchWords() ).toEqual( searchWords );
+	} );
+
+	it( 'should throw an error if the param is not an object', () => {
+		expect( () => getSearchWords( 'lorem' ) ).toThrow( Error );
+	} );
+
+	it( 'should throw an error if the `search` property is not a string', () => {
+		const query = {
+			search: new Object(),
+		};
+
+		expect( () => getSearchWords( query ) ).toThrow( Error );
 	} );
 } );