+++ plugins/woocommerce-admin/packages/navigation/CHANGELOG.md
@@ -1,3 +1,6 @@
+# (unreleased)
+- New method `getSearchWords` that extracts search words given a query object.
+
 # 2.0.0
 
 - Replace `history` export with `getHistory` (allows for lazy-create of history)